requirements.txt
aiohttp==3.11.16
dnspython==2.8.0
tldextract==5.1.2
